requirements.txt
azure-functions==1.12.0
azure-identity==1.12.0
azure-monitor-query==1.1.1
msgraph-core==0.2.2